// src/app/models/category-detail.ts
export interface CategoryDetail {
  bestsellers_date: string;
  books: Array<any>;
  corrections: Array<any>;
  display_name: string;
  list_name: string;
  list_name_encoded: string;
  next_published_date: string;
  normal_list_ends_at: number;
  previous_published_date: string;
  published_date: string;
  published_date_description: string;
  updated: string;
}